+++ src/components/modals/ShopModal.tsx
@@ -8,7 +8,7 @@ import FormRow from "../elements/FormRow.tsx";
 
 export default function ShopModal({ onClose, shop, setShop, onSave, inPlace }: ShopModalInput) {
     const { t } = useTranslation();
-    const coordinateRegexp = /[0-9\.]{3,12} [0-9\.]{3,12}/g;
+    const coordinateRegexp = /^-?[0-9\.]{3,12} -?[0-9\.]{3,12}$/;
 
     const changeType = (e: React.ChangeEvent<HTMLInputElement>, key: string) => {
         const value = e.target.value;